+++ mocks/server.js
@@ -1,5 +1,5 @@
 const statusSuccess = require('../test/fixtures/status.success');
-const suitesSuccess = require('../test/fixtures/suites.success');
+const suitesSuccess = require('../test/fixtures/check_suites.success');
 const repositories = require('../test/fixtures/installation.info');
 const pullRequest = require('../test/fixtures/pull_request.info');
 const user = require('../test/fixtures/token.info');
