+++ test/testWebhookRoute.js
@@ -21,8 +21,8 @@ const pullRequestClosedPayload = require('./fixtures/pull_request.closed');
 const pullRequestMergededPayload = require('./fixtures/pull_request.merged');
 const statusFailurePayload = require('./fixtures/status.failure');
 const statusSuccessPayload = require('./fixtures/status.success');
-const checkSuiteFailurePayload = require('./fixtures/checksuites.failure');
-const checkSuiteSuccessPayload = require('./fixtures/checksuites.success');
+const checkSuiteFailurePayload = require('./fixtures/check_suites.failure');
+const checkSuiteSuccessPayload = require('./fixtures/check_suites.success');
 const pullRequestSynchronizePayload = require('./fixtures/pull_request.synchronize');
 const repositoryCreatedPayload = require('./fixtures/repository.created');
 const repositoryDeletedPayload = require('./fixtures/repository.deleted');
